groups-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.4.0
requests==2.32.2
gunicorn==22.0.0